rename packages -> add a flume.sink
